+++ Changelog.md
@@ -2,6 +2,7 @@
 All notable changes to this project will be documented here.
 
 ## [unreleased]
+- Add Jupyter tester (#284)
 
 ## [v1.10.3]
 - Fix bug where zip archive was unpacked two levels deep instead of just one (#271)